+++ .env.example
@@ -5,3 +5,6 @@ RABBITMQ_HOST="<RabbitMQ Host>"
 RABBITMQ_QUEUE="<RabbitMQ Queue>"
 RABBITMQ_USER="<RabbitMQ User>"
 RABBITMQ_PASS="<RabbitMQ Password>"
+API_URL="<API URL>"
+AZURE_ENDPOINT_TTS= "<Link to Azure TTS>"
+AZURE_ENDPOINT_GPT= "<Link to Azure GPT>"
