+++ tts.py
@@ -18,11 +18,12 @@ async def tts(prompt,tone):
         str: The file path to the saved audio output.
     """
 
-    key = os.getenv("TTS_API_KEY")
+    TTS_API_KEY = os.getenv("TTS_API_KEY")
+    AZURE_ENDPOINT_TTS = os.getenv("AZURE_ENDPOINT_TTS")
     clientTTS = await AsyncAzureOpenAI(
         api_version="2024-12-01-preview",
-        azure_endpoint="https://sebi-m9edwlj5-northcentralus.cognitiveservices.azure.com/",
-        api_key=key,
+        azure_endpoint=AZURE_ENDPOINT_TTS,
+        api_key=TTS_API_KEY,
         
     )
 
@@ -41,33 +42,4 @@ async def tts(prompt,tone):
     return output_audio_path
 
 
-async def transcribe(output_audio_path):
-    """
-    Transcribe an audio file using the Whisper model from OpenAI.
 
-    Given a path to an audio file, this function will transcribe it using the Whisper model
-    from OpenAI and save the transcription to a file named "subtitles.srt" in the same
-    directory.
-
-    Args:
-        output_audio_path: str, the path to the audio file to transcribe
-    """
-    key = os.getenv("OPENAI_API_KEY")
-
-    clientSUB = await AsyncAzureOpenAI(
-        api_version="2024-12-01-preview",
-        azure_endpoint="https://ai-sebimomir-3123.cognitiveservices.azure.com/",
-        api_key=key,
-    )
-
-    result = await clientSUB.audio.transcriptions.create(
-
-        file=open(output_audio_path, "rb"),            
-        model="whisper",
-        response_format="srt"
-    )
-
-    with open("subtitles.srt", 'w', encoding='utf-8') as f:
-        f.write(result)
-    
-    return "subtitles.srt"
